+++ src/Controller/TempController.cpp
@@ -52,7 +52,7 @@ bool TempController::shouldWarm()
 
     const size_t month = now->tm_mon + 1;
     const size_t day = now->tm_mday;
-    const float month_fp = static_cast<float>(month) + static_cast<float>(day) / 30.0;
+    const double month_fp = static_cast<double>(month) + static_cast<double>(day) / 30.0;
     if ((month_fp > 5.5) && (month_fp < 9.0)) {
         // Cool in the months July and August
         return false;
